[source] Add relevant UI content and underlying implementation

+++ webrtc_cpp/src/main/cpp/src/ffi_define_struct.h
@@ -10,8 +10,8 @@
 #include <cstdint>
 
 typedef struct {
-    int64_t width;
-    int64_t height;
+    double width;
+    double height;
     double aspectRatio;
     double frameRate;
     char* facingMode;
